feat: Merge hc_command_util crate into hc_common
