--- libs/hc_command_util/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "hc_command_util"
-version = "0.1.0"
-edition = "2021"
-license = "Apache-2.0"
-publish = false
-
-[dependencies]
-hc_common = { path = "../hc_common" }
-once_cell = "1.10.0"
-regex = "1.5.5"
-
