+++ libs/hc_eslint/Cargo.toml
@@ -7,7 +7,6 @@ publish = false
 
 [dependencies]
 hc_common = { path = "../hc_common" }
-hc_command_util = { path = "../hc_command_util" }
 
 [dev-dependencies]
 tempfile = "3.2.0"
