+++ libs/hc_git_command/Cargo.toml
@@ -6,5 +6,4 @@ license = "Apache-2.0"
 publish = false
 
 [dependencies]
-hc_command_util = { path = "../hc_command_util" }
 hc_common = { path = "../hc_common" }
