+++ libs/hc_git_command/src/lib.rs
@@ -1,8 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
-use hc_command_util::log_git_args;
 use hc_common::context::Context as _;
-use hc_common::{error::Result, hc_error, which};
+use hc_common::{command_util::log_git_args, error::Result, hc_error, which};
 use std::convert::AsRef;
 use std::ffi::OsStr;
 use std::iter::IntoIterator;
