+++ libs/hc_session/Cargo.toml
@@ -10,7 +10,6 @@ dirs = "5.0.1"
 dotenv = "0.15.0"
 hc_analysis = { path = "../hc_analysis" }
 hc_binary = { path = "../hc_binary" }
-hc_command_util = { path = "../hc_command_util" }
 hc_common = { path = "../hc_common" }
 hc_config = { path = "../hc_config" }
 hc_data = { path = "../hc_data" }
